requirements.txt
python-telegram-bot==22.5
APScheduler==3.11.1
pytz==2025.2
pillow==12.0.0
google-genai==1.56.0
python-dotenv==1.2.1
requests
openai>=1.0.0
